Added the seamor driver and UUVApp into the new repository.

+++ scripts/build-local.sh
@@ -22,25 +22,36 @@ export LIB_ABS_DIR=${LIB_ABS_DIR}
 #List here the project in build order. We can automatically enumerate 
 #all the projects but dependecies are not handled. Listing them in order
 #will reduce how many time we need to run this script.
-PROJECT_LIST="labust_tools labust_xml labust_drivers labust_comms labust_navcon labust_uvsim videoray bvt_sdk labust_tools moos_libs"
+UTILS="labust_tools labust_xml labust_drivers labust_comms labust_navcon"
+SIM="labust_uvsim"
+HARDWARE="videoray seamor bvt_sdk labust_bv"
+MOOS="moos_libs"
+PROJECT_LIST="${UTILS} ${SIM} ${HARDWARE} ${MOOS}"
 
 for project in ${PROJECT_LIST}
 do
-	#Remember the source and binary directory.
-	SRC_ABS_DIR=`find ${TRUNK_ABS_DIR} -type d -name ${project}`
-	TARGET_ABS_DIR=${BUILD_ABS_DIR}/${project}
+	#Find the directory, should contain a CMakeLists.txt file.
+	DIRS=`find ${TRUNK_ABS_DIR} -type d -name ${project}`
+	for dir in ${DIRS}
+	do
+		if [ -e ${dir}/CMakeLists.txt ]; then
+			#Remember the source and binary directory.
+			SRC_ABS_DIR=${dir}
+			TARGET_ABS_DIR=${BUILD_ABS_DIR}/${project}
 	
-	#Create the binary directory and position inside it.
-	mkdir -p ${TARGET_ABS_DIR}
-	cd ${TARGET_ABS_DIR}
+			#Create the binary directory and position inside it.
+			mkdir -p ${TARGET_ABS_DIR}
+			cd ${TARGET_ABS_DIR}
 
-	#Call CMake and compile with make.
-	cmake -DCLEAN_BUILD=ON ${SRC_ABS_DIR}
-	make ${*}
+			#Call CMake and compile with make.
+			cmake -DCLEAN_BUILD=ON ${SRC_ABS_DIR}
+			make ${*}
 
-	#Try to collect the *.deb archives.
-	if [ -e ${TARGET_ABS_DIR}/*.deb ] 
-	then
-		cp ${TARGET_ABS_DIR}/*.deb ${PACKAGE_ABS_DIR}
-	fi
+			#Try to collect the *.deb archives.
+			if [ -e ${TARGET_ABS_DIR}/*.deb ] 
+			then
+				cp ${TARGET_ABS_DIR}/*.deb ${PACKAGE_ABS_DIR}
+			fi
+		fi
+	done
 done
